refactor(web): Move request logger to a servlet filter
